:refactor change auth/auth.tsx to auth/index.tsx
