+++ src/app/(auth)/login/page.tsx
@@ -1,14 +1,12 @@
 'use client';
 import React from 'react';
-import Image from 'next/image';
-import { BuildConfig } from '@/config/config';
-import Auth from '@/presentation/components/reuse/auth/auth';
+import Auth from '@/presentation/components/reuse/auth';
 
 class Login extends React.Component<any, {}> {
   render() {
     return (
       <>
-        <Auth  isSignup={false}/>
+        <Auth isSignup={false}/>
       </>
     );
   }
